updated routes
still having errors with displaying data on stats page. Data will display on insomnia, however some of the data is still udnefined and NOTHING is showing up on heroku when I deploy the application. Working on insomnia, not working when deployed to heroku or when loading local file on web browser